Completed demonstator for simple file upload
On running webapp, go to /vod/upload-demog . You will be presented with a upload bar (where you can select the file to upload) and a paginated table of uploads that have taken place. 
The CSV file that will be processed should have the following columns 'first_name', 'last_name' and  'unique_id' to have its data successfully saved to the database.